requirements.txt
altair==5.1.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
chromedriver-autoinstaller==0.6.2
click==8.1.7
exceptiongroup==1.1.3
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.0
outcome==1.2.0
packaging==23.1
pandas==2.1.1
Pillow==9.5.0
protobuf==4.24.3
pyarrow==13.0.0
pycountry==22.3.5
pydeck==0.8.1b0
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
selenium==4.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
spotipy==2.23.0
streamlit==1.27.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
trio==0.22.2
trio-websocket==0.10.4
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
wsproto==1.2.0
zipp==3.17.0
